face_engine/requirements.txt
numpy
torch
torchvision
Pillow
requests
